Patch init to force loading of monolithic sepolicies for Treble roms
The stock init binary is responsible for loading sepolicy. The move to
Treble caused init to load the split sepolicies from /system and
/vendor. However, we had previously been patching the monolithic
sepolicies in / (boot.img). This meant that our sepolicy patches in /
were never being loaded (that is if we do not use Magisk), which
resulted in bind mounts not occurring and causing our wlan.ko module to
not be mounted.

I did some searching in AOSP's system/core repo and found that init
looks for the existence of "/system/etc/selinux/plat_sepolicy.cil"
file to determine if the split sepolicies should be loaded. A further
look in magiskinit sources showed that topjohnwu had been patching the
stock init binary to force the loading of the monolithic sepolicies.

For now we patch the init binary to force the loading of monolithic
sepolicies.

Thanks to topjohnwu for his method of forcing init to load the
monolithic sepolicies, and to osm0sis for his suggestion to patch init
with sed.

Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>